bot/requirements.txt
twitchio==2.9.1
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.3
websockets==12.0
pillow==10.2.0
requests==2.31.0
aiohttp==3.9.1
